refactor(settings): move 3 orchestration sub-components to own files (PR 4a of cluster split)
Pull NotifySettingsSection (with its private helpers OriginToggleRow,
AutoActionTemplateEditor, HandlingRadioGroup, plus the NOTIFY_EVENTS table,
AUTO_ACTION_EVENTS set, and NotifyEventDef interface), PrMonitorSection,
and GuardrailsSection out of SettingsPanel.tsx into dedicated files. Pure
relocation — no behaviour changes, no signature changes. SettingsPanel.tsx
imports them and passes the same props (`orchestrator`, `setOrchestrator`,
`orchProject`, `save`).

This is step 1 of the orchestration cluster decomposition (#4 in the
remaining-tasks plan). Step 2 will extract the parent OrchestrationSection
that uses these three. Splitting the move from the parent extraction keeps
each PR's diff reviewable: this one is mechanical, the next one is
state-shape changes.

- SettingsPanel.tsx: 1489 → 759 lines (−730).
- New files: NotifySettingsSection.tsx (550), PrMonitorSection.tsx (102),
  GuardrailsSection.tsx (89). Net code change: ~+10 lines (just the new
  top-of-file imports + docstrings that did not exist when inline).
- 82/82 settings tests still pass; `pnpm build` clean; biome clean.
